move sighash logic to a separate module
